Shelly/Robot-Selfheal/requirements.txt
# Core dependencies
robotframework>=6.1.1                # Base Robot Framework
robotframework-seleniumlibrary>=6.1.0  # For web UI automation

# Self-healing specific dependencies
beautifulsoup4>=4.12.2              # For HTML parsing in candidate_algo.py
rapidfuzz>=3.3.1                    # For fuzzy string matching in candidate_algo.py
python-dotenv>=1.0.0                # For loading .env files with OpenAI API key
openai>=1.3.5                       # For GPT-4 API integration

# Utility packages
requests>=2.31.0                    # HTTP library for API calls
python-json-logger>=2.0.7           # For structured JSON logging
openpyxl>=3.1.2                    # For Excel file handling (if needed for test data)

# Development dependencies
black>=23.11.0                      # Code formatting
pylint>=3.0.2                       # Code linting
pytest>=7.4.3                       # For unit testing 